requirements.txt
flask
flask-cors
werkzeug
pandas
openai
python-dotenv
openpyxl
gunicorn
psycopg2-binary
SQLAlchemy
